+++ machine/src/activities/index.ts
@@ -4,4 +4,5 @@ export * from './container-activity';
 export * from './fork-activity';
 export * from './interruption-activity';
 export * from './loop-activity';
+export * from './parallel-activity';
 export * from './results';